Fixed formatting issues for ORD-02
